Refactor(component): `NavItem.view`
- Move logic functions to their own module.
- Move `NavItemRouteIndicator.view` to its own module.
